mach-k3: fix decrypted image size
K3 encryption messes with the original binary image by padding it to a
16-byte aligned length and adding a 32 bytes of random data to the end.

The random data is only used to check if the decryption worked.

This length manipulation screws up U-Boot's loader somewhere making the
boot fail. I haven't looked at exactly why it fails, but it's probably
because the DTB is ending up in the wrong place so that subsequent
stages can't find it anymore.

As far as I can tell TI don't encode the original image size in the DER
certificate, so I've hacked the last 4 bytes of the random data to
contain the original image size.